backend/requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
psycopg2-binary
python-dotenv
